Draft: Create an SBOM for local and remote container images
This is a first pass at creating and saving SBOM documents for container
images in a Zarf package.  Currently this persists SBOM documents as
`.zarf-sbom-<IMAGE ID>.json` within the package archive.  We'll probably
want to refine where and how we're storing these but for a quick
proof-of-concept this should do for the moment.

Currently the SBOM is not used or displayed to the user in any way, it
is only being cataloged at time of package creation.  We'll need to make
a pass over the data within the SBOM and ensure it's complete and
accureate as well.